recruiting-mutants/src/main/java/com/mercadolibre/challenge/infrastructure/controller/exception/ControllerExceptionHandler.java
package com.mercadolibre.challenge.infrastructure.controller.exception;

import com.mercadolibre.challenge.core.exception.DNAException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ControllerExceptionHandler {

    @ExceptionHandler(DNAException.class)
    public ResponseEntity<Object> handleHumanException(DNAException exception) {
        return new ResponseEntity(exception.getMessage(), HttpStatus.FORBIDDEN);
    }

    @ExceptionHandler(MethodArgumentNotValidException.class)
    public ResponseEntity<Object> handleValidationExceptions( MethodArgumentNotValidException ex ) {
        return new ResponseEntity(ex.getMessage(), HttpStatus.BAD_REQUEST);
    }

}
